requirements.txt
# ideally pip freeze or pipfile lock or poetry, now simple requirements.txt file
fastapi==0.105.0
pydantic==2.5.2
pydantic-settings==2.1.0
gunicorn==21.2.0
uvicorn[standard]
python-multipart
pypdfium2==4.25.0
pillow==10.1.0
numpy
starlette-context==0.3.6
starlette==0.27.0
psycopg[binary]==3.1.16
psycopg_pool==3.2.0